ACPI: PCI: add Cavium Thunder ECAM
In case of Cavium Thunder platform, we need to use a set of platform
specific ECAM operations for ACPI

Export those from pci-thunder-ecam.c where it is used for OF based
init and use it in pci_gen_host.c.  We use dmi_match functions for
using the platform specific ecam ops.