acpi, pci: Support IO resources when parsing PCI host bridge resources.
Platforms that have memory mapped IO port (such as ARM64) need special
handling for PCI I/O resources. The IO resources on the host bridge has
be registered with pci_register_io_range, and has to be fixed up to be
int the ioport_resource range. The IO resources also need mapping with
pci_remap_iospace.

On release (in case of hot plug removal), we need to call
pci_unmap_iospace to unmap the IO resource.

Signed-off-by: Jayachandran C <jchandra@broadcom.com>
Signed-off-by: Sinan Kaya <okaya@codeaurora.org>
Signed-off-by: Tomasz Nowicki <tn@semihalf.com>