PCI: Handle interaction of ACPI and generic PCI domains
When ACPI is enabled, the ACPI pci segment number is to be used
as the PCI domain.

On ARM64, ACPI and OF can be enabled together, and in that case
we need to use generic domains if ACPI is disabled, and ACPI domains
if ACPI is enabled.

To implement this:
 - Move implementation of pci_domain_nr to arm64 asm/pci.h, and
   return domain number from sysdata in case of ACPI. If ACPI is
   disabled, use the generic bus->domain_nr
 - Update pci_bus_assign_domain_nr() to not setup generic domain
   if ACPI is enabled